+++ src/test/java/BunTests.java
@@ -1,8 +1,8 @@
-package praktikum;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
+import ru.yandex.praktikum.Bun;
 import java.util.Arrays;
 import java.util.Collection;
 import static org.junit.Assert.assertEquals;
